requirements.txt
requests
beautifulsoup4
python-dotenv
coloredlogs
cmake
setuptools
polars
pytest
jupyter
requests-mock